MB-27820: Introduce salt option to cbcollect_info
To make log redaction hashing more secure we're introducing a salt
option to cbcollect_info --log-redaction-salt.
If redaction is enabled but the salt is not provided it will default
to a random uuid.

When logs are collected with redaction enabled we  write the hashed
hexdigest of this salt to a new file log_redaction.salthash.log to
facilitate correlation of salt between log collection runs.

Also updating the --log-redaction option to --log-redaction-level
to make it more explcit.

Change-Id: I66bb01f0f4dcc71877113c8fab8d280d73141e89
Reviewed-on: http://review.couchbase.org/89025
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>