Encode a feature list in DCP replicators' name.
We used to encode a boolean, which indicates whether to enable XAttr
or not, in the name of the dcp replicator process. This was done so
that we can restart all the replicators once the cluster becomes 5.0.
Now we are trying to generalize this scheme so that newer features
can also be enabled in a similar fashion.

The way this is achieved is by encoding the list of features in the
child ID. Whenever a feature is enabled/disabled, the contents of the
list will be different which leads to the restart of downstream DCP
replicator processes.

Change-Id: I43238afabf1068280503ce0dbb09f110d9d68868
Reviewed-on: http://review.couchbase.org/89509
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>