MB-27811: Use INCLUDE_DELETE_TIMES flag while...
...opening dcp connections.

This flag needs to be used only when the cluster has become fully
vulcan. This can be acheived by adding this new feature to the
feature list maintained by the dcp_sup module. That would ensure
that when the cluster turns vulcan all the older connections will
be dropped and newer ones will be established with INCLUDE_DELETE_TIMES
flag set.

Change-Id: I21e72915f9998d05b1fd0c7d6107ed1ebe1e2e1f