[functools] Add a sequence function.
It's sort of a monomorphic version of haskell's
Control.Monad.sequence. Where in haskell it works with any monad, here
it works the way it would work for Either or Maybe monads in
haskell. In simpler words, it applies functions one by one to the
passed argument and collects the results. But it stops short when any
of the functions indicates an error.

Change-Id: I0421e1b957cd6f8b7ecf460667662fb61b0d8c81
Reviewed-on: http://review.couchbase.org/89975
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>