Rework acquire_lease API to accept when_remaining option.
Currently, leader_lease_acquire_worker does lots of smart (and
complicated) things to figure out when to send the request to the
corresponding lease agent. It tries to make sure that no unnecessary
messages are sent, at the same time it tries not to be too late. The
reworked API addresses these considerations in a different way. Once
the worker acquires initial lease, it will send the request to renew
the lease immediately. But it will use the when_remaining parameter to
indicate to the agent not to respond to the request
immediately. Instead, the agent will only respond once the remaining
lease time is lower than when the worker specified via the
when_remaining. This approach is hopefully more resilient to sudden
latency spikes.

Note, this commit only introduces the new parameter. The worker will
be adapted in the following changes. As of this change, everything
should work the way it did before.

Change-Id: I3775717e4691d6d00784d1b5d2c8f640b57a53d7
Reviewed-on: http://review.couchbase.org/89976
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>