Crash acquire worker on unexpected responses.
Because they typically indicate programming error, so we want them to
be caught as early as possible.

Note, that the same does not apply to exceptions. Even though the
exceptions can also be a result of simple programming error, they can
also be true exceptions, e.g., a communication failure or a system
overload. In those situations we want the backoff logic to kick-in. So
we treat the differently.

Change-Id: I67279b1a891692e7c33efa791b8b6aa63f97c517
Reviewed-on: http://review.couchbase.org/90029
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>