MB-28193 Log exit message that terminates service rebalance.
We only expect that to be an exit from the parent, but the ticket
suggests that in some case we might receive an exit message from
somewhere else. From the code inspection, I couldn't find how that
could be, so hopefully, if it's indeed happening, the extra logging
will help.

Change-Id: Ie4da39d892e988fe35d9495d3fb1eff151cc4225
Reviewed-on: http://review.couchbase.org/91424
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ajit Yagaty <ajit.yagaty@couchbase.com>