Slight refactoring of how leases are persisted to disk.
The main change is that when the lease has just been acquired, it's
always persisted with the full requested time period. It's always ok
to honor the lease for a little longer and it clears a way to the next
change that avoids unnecessary writes to disk.

Change-Id: I65129f0334e99c5659c9170a7bf31c699000ed01
Reviewed-on: http://review.couchbase.org/91434
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>