Persist lease to disk only when necessary.
Since the acquirer always requests the same period, the persisted
values are always the same. So it's redundant to persist the same
values again when the lease is simply extended. On a busy system disk
latency can be quite high (especially since we sync the marker
file).

Change-Id: I662883c38a895c478838382615cb8f46631f25d7
Reviewed-on: http://review.couchbase.org/91435
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>