Use {prev_,}acquired_at for lease acquire time estimates.
Before this change, for safety purposes the acquirer always assumed
that the lease period starts the moment the acquire request was sent
to the lease agent. The downside of this is that it's not resilient to
sudden communication latency spikes. With this change, the acquirer
would also use the acquired_at and prev_acquired_at values returned by
the agent when the acquirer already holds the lease. From these two
values the acquirer will know how much time has passed on the agent
from the previous lease acquisition. So now the acquirer has two
estimates for when the new lease period starts: the old one (the time
when the request was sent plus the duration for which the request was
queued on the agent) and the new one (acquier's own previous lease
acquisition time plus the time from previous acquisition according to
the agent). Now the acquirer can just pick the one that is more
recent (the one that extends the most to the future).

It can be seen that if the current error factor between the acquirer's
idea of it and true lease period start is smaller than the
communication latency on extension, then it's better to use
acquired_at+prev_acquired_at estimate. And vice versa. So this way the
acquirer can always maintain the accuracy of the lease period start on
the order of the smallest communication delay it has ever seen.

Change-Id: I8e7b8a1148457a74f02a1c5bea737769646b1f46
Reviewed-on: http://review.couchbase.org/91577
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>