Don't use gen_server timeout for lease acquisition.
Because it's a timeout on the reception of the answer. Sending the
call to the server is not under the timeout. Normally, it's not a big
deal. But here we are talking to a different node, and the send may
block if our communication channel to the other node is backed up.

Change-Id: Iac5da04dc094b07b5c5a793af83a723e748e7f5e
Reviewed-on: http://review.couchbase.org/91580
Reviewed-by: Poonam Dhavale <poonam@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>