MB-28898: Add cpuCount to return of ...
... /pools/default and /nodes/self
This count is the count of logical processors therefore if run
inside a container it will return only those CPU's of the host
visible to the container.
The value will either be an integer or the atom unknown.
unknown indicates one of the following:
1. erlang emulator cannot detect the configured logical processors
2. the node is pre-vulcan and does not support cpuCount

Change-Id: I93342b7d701160d14a67091d74ebebdcf825bfd7
Reviewed-on: http://review.couchbase.org/91730
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Amit Krishnan <krish.amit@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>