Keep an explicit list of nodes participating in quorum in config.
Previously, it was assumed that all active nodes participate in
quorums. It's still mostly true, but instead of using {node, _,
membership} values, we keep an explicit list of quorum nodes in
quorum_nodes config key. Only a master node is allowed to modify this
key. And it also needs to have an agreement from majority of current
and future nodes for this.

There's a dedicated process called leader_quorum_nodes_manager whose
sole responsibility is to manage this key. When the process is started
by the mb_master, it tries to pull the config from all other nodes to
ensure that it has the most up to date version of quorum_nodes.
Similarly, when the key is updated -- it tries to push the config to
the other nodes.

The objective for this change was to try to ensure as much as possible
the all nodes will always use the latest quorum nodes (while still
continuing to use ns_config). Having all the updates go through a
central entity helps in achieving this.

But incidentally this also addresses the following quirk of the
previous implementation. Since it was the node membership that was
used to add/remove nodes to quorums, when nodes were removed from
cluster they were first deactivated and an attempt to sync config was
made. If the config sync failed, the nodes would remain in the cluster
in inactiveFailed state, which would confuse the user.

Change-Id: I59da76c08e860e67a4f26af128f9d966e4ab24f6
Reviewed-on: http://review.couchbase.org/91822
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ajit Yagaty <ajit.yagaty@couchbase.com>