Erlang R16B03 -> 20.3: replace crypto:rand_bytes/1
crypto:rand_bytes/1 is removed in 20th erlang so it should be
replaced with crypto:strong_rand_bytes/1. The problem is that
crypto:strong_rand_bytes/1 can throw a low_entropy exception.
It was recomended by Aliaksey and Artem to not handle this
exception in any way, quote:
"It seems that it should be almost impossible to hit the
 low_entropy exception unless the openssl is misconfigured:
 on Linux openssl should be using /dev/urandom, on windows --
 cryptoapi. And it actually only uses it to seed/reseed its
 own PRNG. Given this, I'd suggest that we just ignored the
 low_entropy exception. It's extremely unlikely that we'll
 ever see it. And it's better not to pretend that we can
 produce secure random bytes when we really cannot."

Links:

https://www.openssl.org/blog/blog/2017/08/12/random/
https://www.openssl.org/docs/manmaster/man7/RAND_DRBG.html

Change-Id: I1c7398e3ee9b1b4de1119c7265935cb5035de14b
Reviewed-on: http://review.couchbase.org/92215
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>