MB-29021 Use previous state to compute new acquire time.
Because that's what the function actually expects. Specifically,
get_prev_acquire_estimate assumes that if have_lease is true, then
lease_acquire_ts is not undefined. But in fact, a semi-updated new
state where have_lease is true (because we just acquired the lease),
but lease_acquire_ts is undefined was passed. If it also happens to be
the case that the agent returned enough information to compute the
time since last acquisition, the worker would crash.

Change-Id: I48c63e6a7a31b208fe37c52126f69b55f6277b74
Reviewed-on: http://review.couchbase.org/92011
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ajit Yagaty <ajit.yagaty@couchbase.com>