unselect roles if the top-level role has beed selected
That is hard to distinguish really selected roles from group
of automatically selected roles in case when Admin or Bucket[*]
role has been selected. This fact is more obvious during further
editing of the user roles. That is because all appropriate wrappers
were marked for highlighting.

I propose not to highlight all appropriate(included) roles
automatically but just disable them. This will help
the users to find the really selected role more faster.

Change-Id: I7ff7adc9c886e0ce07efa2ece4cf10efed659944
Reviewed-on: http://review.couchbase.org/92110
Reviewed-by: Rob Ashcom <rob.ashcom@gmail.com>
Tested-by: Pavel Blagodov <stochmail@gmail.com>