MB-29354 Takeover vbuckets via orchestrator.
Note, this is a customer maintenance patch. The customer in question
is running 5.0 cluster, so this patch does not include the workaround
for the set_vbucket_state to dead race condition (since it doesn't
affect 5.0) and offline config upgrade (since it's a maintenance
patch).

Change-Id: I30e41561f84237995c74925b2009be23847cdded
Reviewed-on: http://review.couchbase.org/93385
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>