Make sure ns_doctor:wait_statuses/2 returns only fresh statuses.
To achieve that we need to exclude "down" statuses from the initial
statuses that we get from ns_doctor.

The only current user of this function is the code for memory quotas
validation. Even though technically this change prevents it from
working in some situations where it previously worked, I believe the
new semantic is clearer. It doesn't pretend that we can deal with down
nodes there. And even previously it only worked only in a limit set of
cases where nodes were down.

Change-Id: I38a5563bf03b6ab2403157ee2dfe1c1c8bcde626
Reviewed-on: http://review.couchbase.org/93656
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>