Add a unit test that exercises offline config upgrades.
It essentially tests that offline upgrades don't crash and that
whatever the end result is, it's the same as the default config.

Note, it used to be the case that when you started clean cluster node,
it would go through offline config upgrades. This was because we
didn't set config_version anywhere in the default config. But I
changed that as part of the change that prevented offline upgrades
from unsupported versions (see http://review.couchbase.org/50447).

Change-Id: Ib2963499318fd5884b78cff361f53e5071cd12c9
Reviewed-on: http://review.couchbase.org/93863
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>