MB-29496 [async] Start trapping exits in the controller later.
Specifically, after registering with parent async and adopters. Even
though it's impossible that they'll be waiting for us to terminate
directly: that would mean that they responded to one of the
aforementioned calls and hence we wouldn't be blocked in the first
place. But it's actually possible that the parent async or an adopter
are waiting on some other process to terminate and that process in
turn is waiting for us to terminate. Since none of these synchronous
calls expect exit signals, that will lead to a deadlock. Not trapping
the exits until later helps to avoid such situation.

Change-Id: I2f216b1cb01217c5782cd28cae5cb029843c8c89
Reviewed-on: http://review.couchbase.org/93970
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ajit Yagaty <ajit.yagaty@couchbase.com>