MB-29732 Abort stuck takeovers with disable_old_master quirk.
Take over may get stuck due to a bug described on the
ticket. Normally, it's undesirable but not fatal, since all the data
remains available. With disable_old_master quirk though, the data
becomes unavailable. So we abort the takeover if it doesn't succeed in
less than 10 seconds. The timeout is deliberately chosen quite
low. But given that there won't be new mutations coming in on the
vbucket, it should be more than enough to complete the takeover under
normal circumstances.

Change-Id: I4501d4fbfef8a61d4a33e019443ffa6bac32c83c
Reviewed-on: http://review.couchbase.org/94482
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>