Use higher resolution clock inside lease agent.
With one millisecond granular timestamps, the difference between two
timestamps can be about 1 millisecond larger/smaller than the true
difference between those moments in time. One millisecond is large
enough duration to cause spurious warnings on the leader about lease
time estimates being in the future.

With this change, leader_lease_agent internally uses native precision
timestamps. The time since the previous acquisition that is reported
to the leader is still in milliseconds. But it will be always rounded
down to the smallest value. So it never overestimates the difference.

Change-Id: I8c2713c616346925f637e3e6a8dabe74bdbdd5e9
Reviewed-on: http://review.couchbase.org/94631
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>