MB-29819 Terminate all replications before disabling old master.
My original thinking was that there was no need to synchronously
terminate replications to the vbuckets that are not in new vbucket
chain since we won't be creating any new replications for the affected
vbucket to or from these nodes. Hence we won't be affected by the
original race (when we disable the old master vbucket, stream_end is
sent to the replica vbuckets; ns_server sees the stream_end and
assumes that replication is gone, but it can actually still exist in
ep-engine for a tiny bit of time). The problem though is that whenever
janitor_agent makes any changes to the replications, it will try to
recreate all replications that are supposed to exist. So if
janitor_agent on one of the old replica nodes receives an unrelated
request before the vbucket move is completed (that's when we clean up
the old replications), it will recreate the replication in
question. Then if it just so happens that the request is received
after ns_server processed the stream_end, but before ep-engine got a
chance to do the same, then the request will fail with "stream already
exists" error. And that will lead to the rebalance failure.

Change-Id: I49bceb9bdbe0985fd885acdbffe05104ecf87944
Reviewed-on: http://review.couchbase.org/94686
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>