Drop /dot/<bucket> and /dotsvg/<bucket> REST endpoints.
In theory, it's convenient to have a way to see all the replications
in a visual form. In practice though, the way it's implemented, it's
barely useful: with 1024 vbuckets it's almost impossible to make out
anything on the resulting graph. I've only ever used it out of
curiosity. So doesn't seem to make a lot of sense to keep it around.

Change-Id: Idb05a5c1b4fc5515c1848e58c6eaaf65c9476dbb
Reviewed-on: http://review.couchbase.org/95155
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>