Move update_bucket handlers into the idle state handler.
This has a side-effect of disallowing bucket updates in the recovery
mode. That's consistent with any other bucket operations that are
handled by the orchestrator.

Unfortunately, new error returned by the orchestrator means crashes on
old nodes. But since the recovery is such little known/used feature,
I'm willing to tolerate this fact for the sake of code simplicity.

Change-Id: I435bbe06f96d7f545eb2fcb5c3da075cad2ec3ee
Reviewed-on: http://review.couchbase.org/95211
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>