Make /controller/stopRebalance safe by default.
It used to be the case that UI would only call
/controller/stopRebalance with onlyIfSafe=1 and then would prompt the
user to confirm the action if needed. This logic was lost though with
the transition to new UI. This change make the "safe" option
default. The parameter name and the response status code is changed in
line with safe/unsafe failover.

Note, that the definition of what's safe and what's not safe is
unchanged. Even if not ideal, it's too late to make any changes on a
vulcan branch.

Change-Id: I1daa48e3561d4d9e15e3dfdc36a859e2dac6800f
Reviewed-on: http://review.couchbase.org/95219
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>