Don't use misc:verify_name/1 in ns_orchestrator.
It was a poor-man's way of dealing with network partitions where,
supposedly, it was possible to have two ns_orchestrator processes
running in the cluster (while one of them not being registered). With
new leader_registry, it's still possible that there'll be two
ns_orchestrator-s registered in the cluster, yet the check will not do
anything useful: each of the ns_orchestrator-s will believe that they
are registered. The undesired effects of this are prevented by the
actual janitor being a leader activity.

Change-Id: I4994079815deee9ad975fc8e87228281a2c9f2ef
Reviewed-on: http://review.couchbase.org/95429
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>