Remove the timeout logic from mb_master:shutdown_master_sup.
It's already the case that the supervisor should terminate in bounded
time. This is achieved through the use of appropriate supervisor
shutdown policies. Having an extra timeout here just negates them. And
in fact, it's not safe to brutally kill supervisors. This may lead to
the supervisor children not being properly terminated.

Change-Id: Ie8098fa00a8b86e1c110590bdfeed6de15627a08
Reviewed-on: http://review.couchbase.org/95433
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>