[bp] Remove the timeout logic from mb_master:shutdown_master_sup.
It's already the case that the supervisor should terminate in bounded
time. This is achieved through the use of appropriate supervisor
shutdown policies. Having an extra timeout here just negates them. And
in fact, it's not safe to brutally kill supervisors. This may lead to
the supervisor children not being properly terminated.

Original-Change-Id: Ie8098fa00a8b86e1c110590bdfeed6de15627a08
Original-Reviewed-on: http://review.couchbase.org/95433
Original-Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Original-Well-Formed: Build Bot <build@couchbase.com>
Original-Reviewed-by: Artem Stemkovski <artem@couchbase.com>

This is in preparation to addressing MB-33750.

Change-Id: I4bd026bf353a9ffdcac4f5b29a8e5209eba2a3b3
Reviewed-on: http://review.couchbase.org/107760
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>