Add git_hooks directory.
Meant to be used in conjunction with git-hooks tool from
https://github.com/icefox/git-hooks (can be installed from homebrew
too).

This is the tool to manage git hooks that should be used by
everybody commiting to the repository. Once, the tool is installed,
it's enough to run git hooks --install.

This patch does not introduce any hooks. It only contains a
pass-through hooks that call hooks that previously existed in the
filesystem. One of those is the hook that repo installs to add
Change-Id to the commit messages.

Change-Id: I0022c3e877786aeb10f11cac91d5ff04241f4a29
Reviewed-on: http://review.couchbase.org/95609
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>