Log samples loader output when loading fails.
It may fail for numerous different reasons. So it's better to just log
the output with the specific reason than to guess what might have gone
wrong. For instance, even if there are indexer/query nodes in the
cluster, it may still fail if storage mode has not been set.

Change-Id: I26965b45a5a28c0fe4121bbd4eadf95a77802527
Reviewed-on: http://review.couchbase.org/95723
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>