Use the new is_localhost API to determine if host...
...address is a local address.

Earlier we were directly using "127.0.0.1" and "::1" addresses to
determine if the node address was local. But the IPv6 localhost
address can be represented in multiple different ways in uncompressed
form. For ex: ::1, 0:0::1 and 0:0:0:0:0:0:0:1 all mean the same.

In scenarios where uncompressed IPv6 localhost addresses are used
to configure the cluster, certain REST APIs and server listing in
the UI behave differently.

The fix for this is to use standard erlang API to parse the address
and then determine if it's a localhost address or not.

Change-Id: Id9efd61bf938d114306bf2776010f8fcdea17eb4
Reviewed-on: http://review.couchbase.org/96018
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>