[goport] Don't close child stdin explicitly.
We already rely on stdout and stderr being closed by the
finalizers (because there's no way to unblock the reader goroutines
without closing the files, but closing the files, in turn, is raceful
since File.Close modifies the File struct as opposed to simply closing
the underlying file descriptor). For simplicity sake, do the same for
stdin.

Change-Id: I17f239656c3a051afec43ac4aa1f17a65fed0c1a
Reviewed-on: http://review.couchbase.org/96202
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>