[goport] Kill try_wait_for_exit function.
Deal with the decoding errors in the main loop instead. If we
encounter one, we send ourselves a message with a delay. The message,
upon reception, will terminate the erlang process. If the supervised
OS process terminates before that, the erlang process will terminate
normally.

Change-Id: Id74f70f291b562ab21722be82b842a333a5d5ffb
Reviewed-on: http://review.couchbase.org/96217
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>