[goport] Defer terminate_port call to terminate function.
When goport:shutdown is called. It makes the behavior more like that
of port_close: the caller is expected to wait for the exit signal to
make sure that the port is dead. Though even there, it can only be
guaranteed if the termination reason is normal.

Change-Id: I279f162a31be054aadfd9b64ee7237bf04394798
Reviewed-on: http://review.couchbase.org/96219
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>