[goport] Don't block reads from the process during writes.
Otherwise, if the supervised process stops reading from stdin when
it's blocked on write to stdout/stderr, this may lead to a
deadlock: currently we always execute one command at the time, so a
blocked write prevents any attempts to ACK more data. My original
thinking was that we'll just make sure that the processes we supervise
are well-behaved ones. But it's easier to just have it work.

A better solution would be to allow concurrent writes and acks, but
it's a bit more evolved. So I might get to it later.

Change-Id: Iacc6ab25ea96dc870772ad492379cc8b723e4056
Reviewed-on: http://review.couchbase.org/96225
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>