MB-30582: Don't pass storageMode as command line...
...argument while starting the indexer service.

The indexer service internally decides when to upgrade the
storageMode from forestdb to plasma. When it updates its
settings in metakv, the ns_server picks up those changes
and restarts the indexer service passing the latest storageMode
value as the command line argument. This caused availability
problems for the indexer service.

The fix here is to not pass the storageMode as a command line
argument while starting the indexer service and the indexer
service bootstraps itself by reading in the indexer settings
from metakv.

Change-Id: Ib527d3c00f72e5eaa1f61bc069c855b1f327c5ac
Reviewed-on: http://review.couchbase.org/97344
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Poonam Dhavale <poonam@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>