MB-30947: Run verify_replication/1 fun as ...
... as part of ns_rebalancer process and not spawn a separate process
to run it.

Running the verify_replication as it currently exists today, i.e., spawn a
separate process to run verify_replication, leads to an intermittent issue where
we can receive an EXIT message of the verify_replication pid when we set
trap_exit flag to true in the ns_rebalancer process (parent to the
verify_replication pid) at a later point in time, which at the time of this
writing occurs in the service rebalancer code.  This issue causes unwanted exit
of rebalance.

Since there is no reason anymore why we can't just call verify_replication fun
directly, we resolve the issue as such.

Change-Id: I3ba88329cbc3a68fe2e23183900c24f0f34cd5ec
Reviewed-on: http://review.couchbase.org/98546
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>