MB-30612: Don't enforce cbas & fts directories...
...to be writable while creating the service startup spec.

Currently, we enforce that cbas and fts service's directories be
writable while preparing the respective service startup spec. If
the directory is not writable then this results in an exception
and this leads to not starting any of the other services.

The fix is to not enforce the "ensure_writable" check. If the
directory is not writable then just log a debug message and
continue.

Change-Id: I1dce3e64c390601a609a220f5106701b1a111c18
Reviewed-on: http://review.couchbase.org/98900
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>