MB-24242, MB-31366: Set relevant vBuckets to ...
... replica state and mark the bucket warmed up after delta recovery.

This prevents the following situations,
1. where autofailover interrupts rebalance incorrectly due to not ready buckets.
It was discovered that we wait for buckets to be warmed in memcached at the
beginning of delta recovery. But, janitor cleanup on the bucket is run only when
rebalance of the bucket starts, which marks the bucket as warmed up. Till the
bucket is warmed up the client traffic is not enabled on the bucket and bucket
is not ready. On a cluster with multiple buckets, the node being recovered will
be considered down by auto-failover logic till rebalance of the last bucket
starts and the last bucket becomes ready.

2. Situation explained in bug MB-31366, i.e, if a delta recovery rebalance is
interupted at a point before a bucket is marked warmed up, while other buckets
have finished rebalance and started serving data, then the delta recovery nodes
are failed over due the bucket that has not been marked warmed up,
leading to potential data loss in the buckets serving data.

Change-Id: I317fbb74a17651b6a057c5970dc411cd914b969a
Reviewed-on: http://review.couchbase.org/99045
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>