MB-31362: Pass listening port range to couchdb VM.
We currently don't pass the listening port range, that is set on
babysitter and ns_server VMs, to the couchdb VM. This can cause
inaccessibility problems to severs started by couchdb VM as the
firewall rules may disallow access to the random ports used.

The fix is to pass down the listening port range to couchdb VM
as well.

Change-Id: Ia193e970c2d30a2af5def89d8bdafda831c0e157
Reviewed-on: http://review.couchbase.org/100072
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>