In case of ephemeral bucket we should ...
.. always check_for_unsafe_nodes in ns_janitor:cleanup.
This is done to prevent the scenario where memcached is restarted on one of the
nodes and somehow the orchestrator doesn't get a chance to run the janitor
cleanup before the rebalance/recovery/flush/failover is initiated, we won't
honor the auto-reprovisioning settings for ephemeral buckets. That is, we'll
simply activate the bucket again on the restarted node and lose the data instead
of promoting the replicas.
Rebalance, recovery and flush may now fail due to this extra check.

Change-Id: I9c5ba9d352903b81602d8398b6797c6994fb4316
Reviewed-on: http://review.couchbase.org/100415
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>