MB-31558: New API to be used by UI to determine...
...if client certificate authentication (CCA) must be employed.

The new endpoint is: /_ui/canUseCertForAuth. This is expected
to be used only by the UI.

This will be the API that UI would call before calling /uilogin
API. The return value from this API will be a JSON blob like
below:

{
    "cert_for_auth": "cannot_use" | "can_use" | "must_use"
}

This API, depending on the CCA settings and whether the request
is carrying a client certificate, one of the values listed above.
The following is how ns_server will decide the value:

* If CCA state is "disable"                -> cannot_use
* If CCA state is "mandatory"              -> must_use
* If CCA state is "enable" and the request -> can_use
  has a client certificate.
* If CCA state is "enable" and the request -> cannot_use
  doesn't have a client certificate.

Change-Id: I24a8448564916c7c75f2991b380e6f61ab0d1070
Reviewed-on: http://review.couchbase.org/100468
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>