MB-30854: Add support for Required IPv4/IPv6 ...
... listeners

+Problem+

Currently memcached attempts to bind IPv4 and IPv6 listening ports
based on the "ipv4" & "ipv6" parameters for each interface listed in
the config file, with the following values:

    true  - Attempt to listen on the given port/protocol.
    false - Don't attempt to listen on the given port/protocol.

If we fail to bind on a given protocol, it is only considered a fatal
error if _no_ protocol could be bound. For example; if IPv4 fails but
IPv6 succeeds then that's treated as a warning and memcached will
continue to startup.

This can be problematic in some deployments - consider an environment
using IPv4, but which has IPv6 present on the system. If memcached
fails to listen on IPv4 then memcached will still start, even though
this node is not accessible.

This is particulary troublesome if the management port (11209) *is*
successful in listening, as ns_server will consider the memcached
process healthy (and potentially distribute vBuckets to it) - even
through clients (using port 11210) cannot connect.

+Solution+

This patch changes the ipv4 & ipv6 interface parameters, adding a new
"required" value which allows ns_server to specify that this
port/protocol *must* bind. There are now three possible values:

* required - The specified protocol must be enabled on this
             interface. If it cannot be enabled, then memcached will
             fail to start.

* optional - The specified protocol should be enabled on this
             interface. If it cannot be enabled, then permit memcached
             to still start.

* off      - Do not attempt to enable the specified protocol on this
             interface.

(The previous boolean values of true and false are still supported for
backwards compatibility; mapping to "optional" and "off"
respectively).

Note: the default value for both ipv4 & ipv6 is "optional" - this is
      /close/ to the previous behaviour - a failure to bind is not
      considered fatal - however it *does* mean that failing to bind
      any protocols on a given interface is no longer fatal. The
      expectation is that ns_server should be specifying at least one
      protocol as "required" going forward.

Change-Id: I70c5a077befc28631656102854c14969abf57eab
Reviewed-on: http://review.couchbase.org/100485
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>