MB-31803: New API to configure networking mode.
Currently, the static_config file is being used to indicate the
networking mode in which the cluster must operate. The problem
with this approach is that static_config file gets overwritten
during an upgrade and the configured networking mode gets lost.

In order to overcome this, a new REST endpoint called
"node/controller/setupNetConfig" has been introduced, through
which the networking mode can be configured. The new API accepts
a parameter called 'afamily' which can either be set to 'ipv4'
or 'ipv6'.

Based on the value of 'afamily', the server either stores
'inet_tcp' or 'inet6_tcp' in a file called 'dist_cfg' under the
/opt/couchbase/var/lib/couchbase/config directory. 'inet_tcp' and
'inet6_tcp' are the values accepted by 'erl' command's 'proto_dist'
parameter. The contents of this file will be read by the init script
in order to pass the right value while starting the Erlang VM.

The following shows the format of the contents written to 'dist_cfg'
file:

<current_dist_type>,<new_dist_type>

The reason for storing both current and new dist types is that
Couchbase server must be stopped & started again for the new conf
to take effect. In order to stop the server, we would need the
distribution type used to start it so that we can connect to the
babysitter VM to perform a graceful shutdown.

Change-Id: Ic67d2f4f22f48780ea800ad53d1eb0b080c6e127
Reviewed-on: http://review.couchbase.org/101203
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>