MB-31912: Change the format used to store config...
...in dist_cfg file.

In the initial implementation, the distribution configuration was
being stored as a string (and a comma separated strings when user
changed the network configuration) in a file. This forced us to
have custom parsing logic at multiple places. To avoid this, we
have decided to store the configuration as Erlang terms so that
we can reply upon builtin APIs to parse the configuration.

Change-Id: I5ec9e0e1d4f43d999b5af32d9e1268dcdeb8e487
Reviewed-on: http://review.couchbase.org/101652
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>