MB-31460: Use lhttpc instead of httpc for HEAD request.
In a HEAD request it is expected that the server does not send a message-body.
The new httpc module waits for the body even on a HEAD request and subsequently
either times out or the socket is closed by the server. A way around the httpc
problem is to use a different client in this case, lhttpc.

Change-Id: I6f1f6f53773add4b2174300f0978430a61dfd380
Reviewed-on: http://review.couchbase.org/102229
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>