Use proto_dist value to determine networking mode.
We currently pass down an erlang environment variable to determine
the networking mode in which we are operating. This was passed down
explicitly to keep the option open to support dual stack mode in
erlang distribution but it seems like this cannot be done on the
latest erlang too. Hence removing the environment variable.

Change-Id: Ie3f6e6421a77d55e705f046af5837451a57ec04d
Reviewed-on: http://review.couchbase.org/102824
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>