Allow IPv6 mode only in enterprise edition.
When we introduced the REST API to configure the IPv6 mode, it
was forgotten to ensure that it was done only in enterprise
edition. This same API will also be used to configure the
cluster encryption failure which is also an enterprise edition
feature.

Change-Id: I6370433331af9a4805889a402682991438cf5921
Reviewed-on: http://review.couchbase.org/102908
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>