MB-31805: Add node UUID to /pools/default output.
We currently only provide the hostname as part of the /pools/default
output. Other services can use the hostname to save some per node
state as we don't allow the hostname to change if the node is already
a part of the multi-node cluster. However, node renaming is allowed
in case of single node cluster. This can create some problems for
the services to update their state. Hence adding the nodeUUID to
the output.

Change-Id: I660526f7e16a977173b93ff5a2b1f20d5c3a8b2b
Reviewed-on: http://review.couchbase.org/103057
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>