Direct cast to ns_rebalance_observer for rebalance events.
Doing a cast to ns_rebalance_observer for interested events instead of
the bunny hop from master_activity_events_ingress to
master_activity_events to cast in ns_rebalance_observer as was done
previously, to prevent delays as much as possible. This is extremely
ugly but effective and easy to maintain.

Change-Id: Icb3ae2df34c3f864b00d8f23223c7e20a8dae546
Reviewed-on: http://review.couchbase.org/103698
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>